+++ test/fixtures/graphics.h
@@ -112,9 +112,9 @@ class Mock2DRenderer : public I2DRenderer, boost::noncopyable {
 public:
     MOCK_METHOD(void, init, (), (override));
     MOCK_METHOD(void, deinit, (), (override));
-    MOCK_METHOD(void, drawImage, (Texture &, const glm::ivec2 &, const glm::ivec2 &, const glm::vec4 &, const glm::mat3x4 &), (override));
+    MOCK_METHOD(void, drawImage, (Texture &, const glm::vec2 &, const glm::vec2 &, const glm::vec4 &, const glm::mat3x4 &), (override));
     MOCK_METHOD(void, drawImage, (Texture &, const glm::mat4 &, const glm::vec4 &, const glm::mat3x4 &), (override));
-    MOCK_METHOD(void, drawRect, (const glm::ivec2 &, const glm::ivec2 &, const glm::vec4 &), (override));
+    MOCK_METHOD(void, drawRect, (const glm::vec2 &, const glm::vec2 &, const glm::vec4 &), (override));
     MOCK_METHOD(void, drawFullTargetImage, (Texture &, const glm::mat3x4 &), (override));
     MOCK_METHOD(void, drawText, (Font &, std::string_view, const glm::vec3 &, const glm::vec3 &, TextGravity), (override));
     MOCK_METHOD(void, withBlendMode, (BlendMode, const std::function<void()> &), (override));
